requirements.txt
absl-py==2.1.0
accelerate==1.4.0
addict==2.4.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
attrs==25.1.0
bitsandbytes==0.45.3
braceexpand==0.1.7
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.12
decorator==4.4.2
deepspeed==0.16.3
diffusers==0.24.0
distlib==0.3.9
dlib==19.24.6
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.1
et_xmlfile==2.0.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
hjson==3.1.0
huggingface-hub==0.25.2
humanfriendly==10.0
hydra-core==1.3.2
icecream==2.1.4
identify==2.6.9
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
iniconfig==2.0.0
insightface==0.7.3
iopath==0.1.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightning-utilities==0.12.0
llvmlite==0.44.0
lmdb==1.6.2
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
nodeenv==1.9.1
numba==0.61.0
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx==1.17.0
onnxruntime-gpu==1.20.1
open_clip_torch==2.30.0
openai-clip==1.0.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
piq==0.8.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
pre_commit==4.1.0
prettytable==3.14.0
proglog==0.1.10
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyiqa==0.1.13
pyparsing==3.2.1
pyrootutils==1.0.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sdata==0.25.0
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
stringzilla==3.11.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==1.0.14
tokenizers==0.15.2
torchdata==0.9.0
torchdiffeq==0.2.5
torchmetrics==1.6.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
virtualenv==20.29.3
wandb==0.19.6
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
xformers==0.0.28
xlrd==2.0.1
XlsxWriter==3.2.2
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
decord==0.6.0
soundfile==0.13.1
audiomentations==0.39.0
librosa==0.11.0
numpy-minmax==0.4.0
numpy_rms==0.5.0
av==12.0.0
cffi==1.17.1
pycparser==2.22
python_stretch==0.3.1
clint==0.5.1